Rename to sakeejsonrpc
