+++ __init__.py
@@ -1,4 +1,4 @@
 # SPDX-License-Identifier: GPL-3.0
 
 __all__ = ["xbmc", "xbmcgui", "xbmcaddon", "xbmcgui", "colors", "inputstreamhelper",
-           "stub", "xbmcplugin"]
+           "stub", "xbmcplugin", "sakeejsonrpc"]
